+++ components/hosted_media/src/static/video.js
